+++ app/pages/PuzzleView.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import {Text, View} from 'react-native';
+
+const PuzzleView = () => {
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}>
+      <Text>Puzzle Question </Text>
+    </View>
+  );
+};
+export default PuzzleView;
