+++ app/pages/PuzzleWritingPhoto.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import {Text, View} from 'react-native';
+
+const PuzzleWritingPhoto = () => {
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}>
+      <Text>Puzzle Photo </Text>
+    </View>
+  );
+};
+export default PuzzleWritingPhoto;
