+++ app/pages/PuzzleWritingVoice.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import {Text, View} from 'react-native';
+
+const PuzzleWritingVoice = () => {
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}>
+      <Text>Puzzle Voice </Text>
+    </View>
+  );
+};
+export default PuzzleWritingVoice;
